+++ tests/fixtures/acquiring/employees.ts
@@ -1,0 +1,9 @@
+export const acquiringEmployeeListFixture = {
+  list: [
+    {
+      extRef: "abra_kadabra",
+      id: "3QFX7e7mZfo3R",
+      name: "Артур Дент",
+    },
+  ],
+} as const;